refactor testing into different python files for analysis_manager
